Sort section images

+++ app/sections/sql/sections.sql
@@ -1,11 +1,11 @@
 SELECT
   s.id::text AS section,
-  s.start,
-  s.end,
+  s.start::float,
+  s.end::float,
   coalesce(
     coalesce(s.offset_compact,s.offset),
     0
-  ) AS offset,
+  )::float AS offset,
   s.location
 FROM section.section s
 JOIN section.locality l ON s.location = l.name
